sort 100

+++ check/ft_check_sorted.c
@@ -6,13 +6,13 @@
 /*   By: mayache- <mayache-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/02 22:16:41 by mayache-          #+#    #+#             */
-/*   Updated: 2023/04/02 22:17:30 by mayache-         ###   ########.fr       */
+/*   Updated: 2023/04/10 20:59:51 by mayache-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../push_swap.h"
 
-int ft_sorted(long long *arr, int size)
+int ft_sorted(long *arr, int size)
 {
 	int i = 0;
 	while (i < size - 1)
